+++ lib/search_solr_tools/harvesters/base.rb
@@ -50,7 +50,7 @@ module SearchSolrTools
 
         # Some docs will cause solr to time out during the POST
         begin
-          RestClient.get(url) do |response, _request, _result|
+          RestClient.get(url, verify_ssl: OpenSSL::SSL::VERIFY_NONE) do |response, _request, _result|
             success = [200, 204].include?(response.code)
             logger.error "Error in ping request: #{response.body}" unless success
           end
@@ -145,7 +145,7 @@ module SearchSolrTools
 
         # Some docs will cause solr to time out during the POST
         begin
-          RestClient.post(url, doc_serialized, content_type:) do |response, _request, _result|
+          RestClient.post(url, doc_serialized, content_type:, verify_ssl: OpenSSL::SSL::VERIFY_NONE) do |response, _request, _result|
             success = [200, 204].include?(response.code)
             unless success
               logger.error "Error for #{doc_serialized}\n\n response: #{response.body}"